Setup jest for unit testing
