+++ index.html
@@ -56,7 +56,10 @@
   </div>
 
   <script src="./js/localbase.min.js"></script>
+  
   <script src="./js/init.js"></script>
+  <script src="./js/file-upload.js"></script>
+  <script src="./js/slider.js"></script>
   <script src="./js/canvas.js"></script>
 </body>
 </html>
